Revert "Converted User.js -> User.ts"
This reverts commit 4dc2e889467f1d64e1b8d627419a44c57101bfea.
